requirements.txt
Flask==3.0.3
redis==5.2.0
pillow==11.0.0

